Expose IDs in REST services
They are useful for list rendering keys and to reference for
update/delete. The reason Spring doesn't expose them by default is that
best practice is to keep real database keys hidden and expose GUIDs to
the client, but this is just a private application so the complexity of
that solution isn't worth it.